cleaned up appjsx
